Extract AnteHandler from Feeburn module
